refactor: 3단계 - 자동차 경주 - 자동차 전진, 후진 기능 테스트 이동 및 중복 제거
